* Refactor some code. * Calculate the One Rep per workout. * Group workouts by type and set the max rep for each one.
