+++ app/src/main/java/com/tanishkej/onerep/data/model/WorkoutGroups.kt
@@ -1,0 +1,7 @@
+package com.tanishkej.onerep.data.model
+
+data class WorkoutGroups(
+    val maxOneRep: Int,
+    val workoutType: String,
+    val workouts: List<Workout>
+)
