+++ app/src/main/java/com/tanishkej/onerep/ui/workouts/states/WorkoutUiState.kt
@@ -1,9 +1,9 @@
 package com.tanishkej.onerep.ui.workouts.states
 
-import com.tanishkej.onerep.data.model.Workout
+import com.tanishkej.onerep.data.model.WorkoutGroups
 
 sealed interface WorkoutUiState {
-    data class Success(val workouts: List<Workout>) : WorkoutUiState
+    data class Success(val workoutGroups: List<WorkoutGroups>) : WorkoutUiState
     data object Error : WorkoutUiState
     data object Loading : WorkoutUiState
 }
